requirements.txt
nyct-gtfs>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0

# Development / linting
flake8>=7.0.0
mypy>=1.8.0
pytest>=8.0.0
